requirements.txt
django==3.2.8
psycopg2-binary==2.9.1
django-simple-history==3.0.0
djangorestframework==3.12.4

